move business profile to different directory
