+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
     "plugins":[
         "react",
-        "jsx-ally",
+        "eslint-plugin-jsx-a11y",
         "import"
     ],
     "extends": "airbnb"
