+++ src/unittestExample/sum.js
@@ -1,5 +1,6 @@
-// @flow 
+//@flow
 function sum(a: number, b: number) {
     return a + b;
-}
+}gfg
+
 export default sum;
